Feature/tc 135

+++ src/main/java/codeenthusiast/TrainingCenterApp/muscle/MuscleServiceImpl.java
@@ -1,37 +1,81 @@
 package codeenthusiast.TrainingCenterApp.muscle;
 
-import codeenthusiast.TrainingCenterApp.abstracts.AbstractDTO;
 import codeenthusiast.TrainingCenterApp.exceptions.EntityAlreadyExistsException;
 import codeenthusiast.TrainingCenterApp.exceptions.EntityNotFoundException;
-import codeenthusiast.TrainingCenterApp.image.ImageDTO;
+import codeenthusiast.TrainingCenterApp.image.ImageServiceImpl;
+
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class MuscleServiceImpl implements MuscleService {
 
     private final MuscleRepository muscleRepository;
 
+    private final ImageServiceImpl imageService;
+
     private final MuscleMapper muscleMapper;
 
-    public MuscleServiceImpl(MuscleRepository muscleRepository, MuscleMapper muscleMapper) {
+    public MuscleServiceImpl(MuscleRepository muscleRepository, ImageServiceImpl imageService, MuscleMapper muscleMapper) {
         this.muscleRepository = muscleRepository;
+        this.imageService = imageService;
         this.muscleMapper = muscleMapper;
     }
 
     @Override
-    public void deleteById(Long id) {
-        checkExistence(id);
-        muscleRepository.deleteById(id);
+    public MuscleDTO findById(Long id) {
+        Muscle muscle = findEntityById(id);
+        return muscleMapper.mapToDTO(muscle);
     }
 
     @Override
-    public void checkExistence(Long id) {
+    public Muscle findEntityById(Long id) {
+        return muscleRepository.findById(id).orElseThrow(
+                () -> new EntityNotFoundException(id));
+    }
+
+    @Override
+    public List<MuscleDTO> findByMovementId(Long id) {
+        List<Muscle> muscleList = muscleRepository.findByMovementId(id);
+        return muscleMapper.mapToDTOs(muscleList);
+    }
+
+    @Override
+    public MuscleDTO update(Long id, MuscleDTO dto) {
+        this.existsById(id);
+        dto.setId(id);
+        Muscle updatedMuscle = muscleMapper.mapToEntity(dto);
+        return save(updatedMuscle);
+    }
+
+    @Override
+    public MuscleDTO create(MuscleDTO dto) {
+        checkExistenceByName(dto.getName());
+        Muscle muscle = new Muscle(dto);
+        return save(muscle);
+    }
+
+    @Override
+    public MuscleDTO save(Muscle muscle) {
+        Muscle savedMuscle = muscleRepository.save(muscle);
+        return muscleMapper.mapToDTO(savedMuscle);
+    }
+
+    @Override
+    public void deleteById(Long id) {
+        if (existsById(id)) {
+            muscleRepository.deleteById(id);
+        }
+    }
+
+    @Override
+    public boolean existsById(Long id) {
         if (!muscleRepository.existsById(id)) {
             throw new EntityNotFoundException(id);
         }
+        return true;
     }
 
     @Override
@@ -43,61 +87,17 @@ public class MuscleServiceImpl implements MuscleService {
     }
 
     @Override
-    public MuscleDTO findById(Long id) {
-        Muscle muscle = muscleRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException(id));
+    public MuscleDTO addImage(Long id, MultipartFile file) {
+        Muscle muscle = findEntityById(id);
+        imageService.createNewImage(file, muscle);
         return muscleMapper.mapToDTO(muscle);
     }
 
     @Override
-    public List<MuscleDTO> findByMovementId(Long id) {
-        List<Muscle> muscleList = muscleRepository.findByMovementId(id);
-        return muscleMapper.mapToDTOs(muscleList);
-    }
+    public MuscleDTO removeAllImages(Long id) {
+        imageService.deleteImagesByMuscleId(id);
+        return findById(id);
 
-    @Override
-    public MuscleDTO update(Long id, MuscleDTO dto) {
-        checkExistence(id);
-        dto.setId(id);
-        return save(dto);
-    }
-
-    @Override
-    public MuscleDTO create(MuscleDTO dto) {
-        checkExistenceByName(dto.getName());
-        Muscle muscle = new Muscle(dto);
-        return save(muscleMapper.mapToDTO(muscle));
-    }
-
-    @Override
-    public MuscleDTO save(MuscleDTO dto) {
-        Muscle muscle = muscleMapper.mapToEntity(dto);
-
-        return muscleMapper.mapToDTO(muscleRepository.save(muscle));
-    }
-
-    @Override
-    public MuscleDTO addImage(Long id, ImageDTO image) {
-        MuscleDTO muscle = findById(id);
-        muscle.getImages().add(image);
-        return save(muscle);
-    }
-
-    @Override
-    public List<Long> removeAllImages(Long id) {
-        MuscleDTO muscle = findById(id);
-        List<Long> idList = getIdOfDeletedImages(muscle.getImages());
-        muscle.getImages().clear();
-        save(muscle);
-
-        return idList;
-    }
-
-    @Override
-    public List<Long> getIdOfDeletedImages(List<ImageDTO> images) {
-        return images.stream()
-                .map(AbstractDTO::getId)
-                .collect(Collectors.toList());
     }
 
 }
